Adding Snap-on Logo
Added snapon_logo.svg to source images and updated json to add the logo to the partners page. Also updated <a> tag css to style width. Updated a few of the logos on the page to balance size of the logos a bit better so the square logos don't render significantly larger in emphasis than wide logos.